Refactor script
